[FEAT]拆 user 裡動態載入表單的部分
